+++ apps/site/assets/css/_icons.scss
@@ -294,6 +294,7 @@ a {
   color: $brand-commuter-rail;
   padding-left: $base-spacing / 4;
   padding-right: $base-spacing / 4;
+  white-space: nowrap;
 }
 
 .c-icon__bus-pill {
